+++ src/sequence/SequenceError.ts
@@ -1,0 +1,5 @@
+import { CollectionError } from '../collection/CollectionError.js';
+
+export class SequenceError extends CollectionError {
+  // NOOP
+}